refactor: update voting to new approach
