+++ koala/cogs/__init__.py
@@ -7,3 +7,4 @@ from .intro_cog import IntroCog
 from .react_for_role import ReactForRole
 from .text_filter import TextFilter
 from .verification import Verification
+from .voting import Voting
