+++ koala/cogs/voting/__init__.py
@@ -1,0 +1,2 @@
+from . import utils, db, log, models
+from .cog import Voting
